Fix the Integer type of the aircraftId to Long

+++ src/com/ubei/http/entity/Flight.java
@@ -10,6 +10,6 @@ public class Flight {
     private String departureAirportCode;
     private LocalDateTime arrivalDate;
     private String arrivalAirportCode;
-    private Integer aircraftId;
+    private Long aircraftId;
     private FlightStatus status;
 }
